Finished implementing delete user
