+++ Backend/src/main/java/tech/mineapp/security/ApplicationSecurity.java
@@ -8,15 +8,15 @@ import org.springframework.security.config.annotation.web.configuration.WebSecur
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-import tech.mineapp.service.UsersService;
+import tech.mineapp.service.UsersServiceImpl;
 
 @EnableWebSecurity
 public class ApplicationSecurity extends WebSecurityConfigurerAdapter {
 	
-	private final UsersService userService;
+	private final UsersServiceImpl userService;
 	private final PasswordEncoder passwordEncoder;
 	
-	public ApplicationSecurity(UsersService userService, PasswordEncoder passwordEncoder) {
+	public ApplicationSecurity(UsersServiceImpl userService, PasswordEncoder passwordEncoder) {
 		this.userService = userService;
 		this.passwordEncoder = passwordEncoder;
 	}
